+++ tf-modules/ingest/iam.tf
@@ -76,6 +76,7 @@ data "aws_iam_policy_document" "step_policy" {
       "iam:GetPolicy",
       "iam:GetPolicyVersion",
       "iam:GetRole",
+      "iam:PassRole"
     ]
     resources = ["*"]
   }